Add Etsy sponsorship to the Inspector page
It's important that we recognize the support that Etsy
has offered for the Inspector. It's obviously an open source project
and many more engineers will contribute, but there's no denying that
this project is where it's at now because of Etsy.

Etsy will be doing a blog post about the Inspector sometime this month
and at that point we can link to that from the a tag.