Fix initial viewport jump in chrome.
Prior to this fix, the page would load in an odd fashion on mobile
chrome. There would be a viewport width jump. By setting the
initial-scale to 1 this problem is magically fixed.

I am leaning on the side of this is a bug in chrome, however I am not
sure where the fault lies.