Fix flash of page content
Move links to the header - the critical task will convert links
on the home page and inspector page - on other pages we still should keep links
to stylesheets in the header to avoid flash of unstyled content (fouc).
The grunt-critical task has now common configuration shared between targets and we can now build targets separately if required.
This probably is similar to what was already mentioned in #288 PR.
- update Grunt critical task to include Inspector page
- all stylesheets in the same logic block following Jade base practises
- all stylesheets in the same order in all templates: our code then fonts