Remove *primary image of page* property
This is cleanup commit required to be done before adding tags to Inspector
page.
This removed property is not picked by machines for some reasons,
probably due to use of svg. That is something to be investigated in future.
The commit remove this (and related) property to not use them on Inspector page
- which, as application - has no primary image of page meta.