Add software application schema to Inspector page
This commit brings the same tags as used to describe Inspector on Mj landing page.
The difference is that here the entire page is about Inspector
and its related content.
This commit changes description text to one used on Inspector section on the
landing page, which is longer and more enjoyable to read I think.