:bug: Enhance rendering of top bar markup. Fixes #444
This commit adds some little changes to markup to at
least partially fix #444:
- the critical task does not take into account hidden,
not rendered content - for this reason styling for search
input/bar are not added to critical CSS and unstyled content
impact rendering of top bar during page load. To elevate this
problem an inline style is added to .searchbar selector and
removed via JavaScript related code
- all *search* related assets are now loaded > after main assets
style import is moved and JS inclusion is moved as the content is
not required at all when page is started
- the embedded navbar styles are not applied by critical task. Inline
styling is added to UL element to remove default white space rendered
around top navigation to elevate rendering problem.

Thanks!